requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.23.1
diffusers==0.11.1
numpy==1.24.3
scipy==1.10.1
scikit-image==0.20.0
opencv-python==4.8.0.76
Pillow==10.0.0
pywavelets==1.4.1
tqdm==4.65.0
